+++ libsolidity/interface/CompilerStack.cpp
@@ -300,7 +300,7 @@ bool CompilerStack::analyze()
 
 		if (noErrors)
 		{
-			SMTChecker smtChecker(m_errorReporter, m_smtlib2Responses);
+			SMTChecker smtChecker(m_errorReporter, m_smtlib2Responses, m_smtSolver);
 			for (Source const* source: m_sourceOrder)
 				smtChecker.analyze(*source->ast, source->scanner);
 			m_unhandledSMTLib2Queries += smtChecker.unhandledQueries();